Update multiple instance configuration. Instance names are auto-generated.
With this commit, 'PERIPHERALS' is a list that contains corenames of the
peripherals and the amount of times a specific corename appears is equal
to the number of instances desired.

For example:
PERIPHERALS ?= UART UART REGFILEIF UART
Would have 3 uart instances named 'UART0', 'UART1', and 'UART2' and 1 regfileif instance named 'REGFILEIF0'.

Foldernames of peripherals are not used in this implementation.