Enable portmap connections between peripherals in same system. Portmap now allows connections between peripherals of the same system, for example: SUT uart instance 0 with SUT uart instance 1.
SUT can only be build with configured portmap (Unlike previously, where it could be
built without a portmap).

Changed scripts to create SUT as if it were a Tester when running with a
'core under test'. These changes, allow for configuring connections in
the peripheral_portmap with the name 'Tester' instead of 'SUT' when
building with a CUT.

Moved peripheral_portmap.conf to ROOT_DIR.

Refactored some python code.