Fix default/reset value for rstrb. Rstrb contains the valid bytes of a 4 byte word at any given time. When TDATA_W is 1 byte, then rstrb starts at 0000, and follows sequence 0001, 0011, 0111, 1111, (next word) 0001, 0011, ... When TDATA_W is 2 byte, then rstrb starts at 0000, and follows sequence 0011, 1111, (next word) 0011, 1111, ... When TDATA_W is 4 byte, then rstrb starts at 0000, and follows sequence 1111, (next word) 1111, (next word) 1111, ... (With TDATA_W = 4 byte, all bytes are always valid since it reads one full word each time). When a reset signal is activated, the number of bytes valid should be zero.
This commit fixes a bug where the reset value of rstrb was always 0001.
Now the correct (default) value is used according to TDATA_W when the rstrb overflows (goes to next word).
The correct value of zero is used when the rstrb is reset due to a reset signal (no bytes should be valid).