Revert GET_DIRS back to version previous to merge. The GET_DIRS version of iob-soc/system gets the value of a variable without expanding macros and replaces every $(ROOT_DIR) by a locally defined $(ROOT_DIR_TMP). However, some variables contain macros that depend on $(ROOT_DIR) but since these macros are not expanded, the $(ROOT_DIR) is not replaced.
To fix this, we will use the version of GET_DIRS that was previously on
the Tester (before the merge of iob-soc/system). This version overrides
$(ROOT_DIR) variable temporarily and expands the macros of the variables
with this new temporary $(ROOT_DIR) value. This version has the
advantage that it replaces all $(ROOT_DIR) by the new value, but it may
not be able to override it in some cases, such as, when the ROOT_DIR is
passed as an argument to the makefile target (this would not allow the
ROOT_DIR variable to be overriden again).