feat(peripheral_portmap): Allow portmap with `internal` system interface. This commits adds the new `internal` keyword to the peripheral portmap scripts, allowing peripheral ports to be mappend to an internal system wire.
Previously, it was only possible to map a peripheral port to either
another peripheral port or to an IO of the system (external interface).
Now, besides these, the port can also be mapped to an internal wire of
the system, allowing it to connect to other (manually defined) system components that are
not handled by these scripts.

The peripheral portmap consists of a tuple of two items, each of them
containing a corename. When mapping a peripheral to another interface,
the possible corenames for that interface are:
- `internal`: Keyword reserved to map to an internal system wire. The
  name of the wire is based on the value given in the `if_name` field.
- `external`: Keyword reserved to map to a system IO. The
  name of the IO is based on the value given in the `if_name` field.
- <peripheral corname>: The name of another peripheral to map the
  current port to.