More preparations for BOOT as submodule
ROM hex almost ready for use, another SWReg added to BOOT to control the
CPU reset, preboot may be ready(?), iob_soc_boot.v has the full ROM
inside with wires going in and out (in is CPU req, out are CPU resp and
ROM data, and also the CTR register output, and in the future the RST
one too).

The code does NOT compile though. It's not ready. Something is wrong.