CPU is reading and processing preboot correctly
The CPU reads the preboot instructions, all correctly, and resets in the
end. After that, problem below.

Now the problem is MEM which... doesn't work :). xxxx on rdata for some
reason... So not sure if it actually copies anything to memory or only
attempts to, and the bootloader never gets to start because it's all
xxxx anyway.

Also update LIB and PICORV32.