Revert "[LIB] updated submodule"
This reverts commit 0cd867447ace07f7a8a50678af88c9970169b865.

The `axi_m_port` interface of iob-soc is a bus with multiple `axi_m`
interfaces. Therefore, we should specify that the `ext_mem` component of
iob-soc only uses the first interface of that bus.
Without the bit select (even for 1 bit signals), the Verilog will assign
the entire bus to the `ext_mem` component.