feat(copy_srcs): Re-implement functionality of copy_srcs() function. New copy_srcs function now copies all files from the setup directory of a module, without copying files from its superclasses. This new function will rename every string found inside files and in their filenames, that matches the name of the modules class. The new name given to every string found will be the one given in 'self.name'. So, if we setup an IOb-SoC module like so `iob_soc(name="iob_soc_sut")` then the `iob_soc.v` file from the setup dir will be renamed to `iob_soc_sut.v` in the build dir.
Current implementation may throw errors when copying and trying to
rename binary files.

Note: We could use the following line of code to get the core's setup
dir:
`sys.modules[self.__module__].__file__`
However, with the current implementation it does not work because we are currently
using python modules from an external folder (defined by PYTHONPATH) (not in the setup dir).