refactor(iob_soc.py): Use `create_*` methods to generate wires, ports and groups. As discussed privately, the wires and ports are now generated by using dedicated `create_wire` and `create_port` methods. - The `create_wire` method creates a wire, and appends it to the project   global list of wires named `global_wire_group_list`. (If the wire does   not have a group, one will be created automatically.)   (I also made groups for wires to be consistent with the ports, and   because they are useful to make grouped connections of related wires, like the ones from   rs232. With groups, we can connect the entire rs232 group to the uart   in one line, instead of connecting each wire, one by one.) - The `create_port` method creates a port, and appends it to the local   module's port list named `port_group_list`.
I have also created similar functions for grouping wires and ports,
named `create_wire_group` and `create_port_group`.
- The `create_wire_group` method creates a group of wires and appends it to
  the project global list of wires named `global_wire_group_list`.
- The `create_port_group` method creates a group of ports and appends it to
  the module's port list named `port_group_list`.

These grouping functions, receive an argument named `wires` (or `ports`), to allow the user to make a custom (non standard) group.
I am not sure if this is the best approach, or if we should always force the user
to use the `create_wire`/`create_port` methods.
If this argument is not provided or is empty, the function will create
the wires/ports automatically, based on the group name.
Example: The "rs232" group name will create the wires/groups of that
standard interface automatically.

An alternative way to assign a wire to a group is by using the argument
`group` of the `create_wire`/`create_port` methods.

To connect a port to a wire/wire_group, use the `connect` argument of the
`create_port`/`create_port_group` methods.
This argument receives a string with the name of the global wire/wire_group to
connect to.

When instantianting a module, to connect its ports/port_groups to the
wires/wire_groups, use the `connect` argument.
This argument is similar to the one used in the `create_port`/`create_port_group`.

This commit is a work in progress!
- I am not sure if the groups are implemented in a way that is intuitive enough for the user.
- We should discuss how to extract individual wires from groups. For now
  I have used the `get_wire_from_group(<group_name>, <wire_name>)` function
- We should discuss what datatypes the `connect` argument should receive.
  For example, if it receives a string, should it assume it is a wire
  name or a wire_group name?
- We should also discuss if wires and wire_groups should have unique
  names, since they are global (accessible by any component in the
  entire project)
  Does a wire that is in a group need to have a unique name? Does it
  need to be different from a wire that is in a different group?

In the iob_soc.py file, I have only changed the port section and the start of the iob_soc module description (the part where we instance the cpu).
The rest of the module is unchanged keep this commit simple.