feat(python): Replace dictionaries by objects - Replace dictionaries of the 'signals' list by objects of the class   'iob_signal'. - Replace dictionaries of the 'csrs' list by objects of the class   'iob_csr'. - Update python scripts to support the new objects.
NOTE: The user still passes dictionaries in the lists of the
`create_wire`, `create_port`, `create_csr_group` methods.
However, these dictionaries will be automatically converted into objects for internal use in our python scripts.
In the future, if we decide that the user should pass objects instead of
dictionaries, then the modification should be as simple as removing the
`convert_dict2obj_list()` call inside the `create_*` methods.