feat(nix): Add py2hwsw to nix [WIP] Add py2hwsw to Nix. However, there is currently an issue, because nix removes the write permission from all files inside the py2hwsw directory (located in the /nix/store/), therefore, the files copied to the build dir will also be missing the write permission, and that will cause setup errors.
I have also tried to use the `fetchgit` instead of `fetchFromGitHub`,
but it doesn't seems to have the same problem.

From what I understand, nix removes the write permission from all files
inside the nix store, because those files are supposed to be immutable.
In this py2hwsw case, we are not trying to modify those files, however
we still need to have the correct permissions for the generated build
directory.

We could probably update the py2hwsw scripts to ensure that the write
permission is set on all files, but this would not be a good approach.